refactor: Extract 'Sync' from DeckPicker
* extracted createSyncListener + dependencies

Small start: eventually remove 'DeckPicker' as a dependency
